Fixes the converter to ignore the .schema for valid .mcd
.schema files are temporary files that are generated by the cytof
software during acquisition. They get deleted once the .mcd writing
is finished. They should only be present if the .mcd writing failed
in which case they can be used to rescue the corrupted .mcd.

Unfortunately - due to some backup schemes, the .schema file can also
be backed up before being deleted, causing the .schema file still being
present.

Convertfolder2imcfolder until now allways used the .schema
file for the metadata when present, which can lead to problems as the
backuped schema files are usually not the latest ones. Thus I changed
the code to attmept to read the .mcd without the .schema and only
attempt to consider the .schema if the .mcd cannot be decoded without
it.