Make 'bigtiff=True' default for ome mode
This changes the default mode for tiffs to 'bigtiff' for 'ome'
tiffs to True  and leaves it False for ImageJ tiffs.
Thus OME tiffs are (per default) consistently written as bigtiffs,
which is important as sometimes the raw data files can be >4gb.

ImageJ compatible tiffs (often used for analysis) are consistently
written as not bigtiffs, which is reasonable, as ImageJ compatibility
is potentially incompatible with bigtiffs (see warning in tifffile
source code "writing incompatible BigTIFF ImageJ")